Preliminary final version.
Might do changes over the weekend as I am unable to test the functionality in it's entirety due to the database being down.

Been up for 20 hours so need to get some sleep, if something doesn't work then it'll be fixed before the presentation.

In the end it's pretty much became the Network assignment, I really wanted to get the fragments and recycler working, but oh well. Maybe over the weekend.